Allow to complete a job via ZBClient
